requirements.txt
pandas>=2.0
openpyxl>=3.1
selenium>=4.10
python-dotenv>=1.0
beautifulsoup4>=4.12